temp/mongo_phi_masker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.9
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
black==23.9.1
bleach==6.2.0
bokeh==3.7.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.12
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
dill==0.3.9
distributed==2025.2.0
dnspython==2.4.2
exceptiongroup==1.2.2
fastjsonschema==2.21.1
flake8==6.1.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isoduration==20.11.0
isort==5.12.0
Jinja2==3.1.6
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
locket==1.0.0
lz4==4.4.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mistune==3.1.3
msgpack==1.1.0
multidict==6.2.0
multiprocessing-logging==0.3.4
mypy==1.5.1
mypy-extensions==1.0.0
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.26.0
overrides==7.7.0
packaging==24.2
pandas==2.1.1
pandocfilters==1.5.1
partd==1.4.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
propcache==0.3.0
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==19.0.1
pycodestyle==2.11.1
pycparser==2.22
pydeps==3.0.1
pydyf==0.11.0
pyflakes==3.1.0
Pygments==2.19.1
pylint==3.3.5
pymongo==4.5.0
pymongo-inmemory==0.2.11
pyparsing==3.2.1
pyphen==0.17.2
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==26.3.0
referencing==0.36.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
secure-smtplib==0.1.1
Send2Trash==1.8.3
simpervisor==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stdlib-list==0.11.1
tblib==3.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
weasyprint==64.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xyzservices==2025.1.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
zopfli==0.2.3.post1
